Just move middleware files and update rwjs/vite import paths
