+++ packages/vite/src/middleware/index.ts
@@ -1,3 +1,1 @@
-export * from './MiddlewareRequest.js'
-export * from './MiddlewareResponse.js'
 export * from './types.js'
