+++ packages/vite/src/middleware/register.ts
@@ -3,12 +3,12 @@ import type Router from 'find-my-way'
 import type { ViteDevServer } from 'vite'
 
 import { getPaths } from '@redwoodjs/project-config'
+import type { MiddlewareRequest } from '@redwoodjs/web/dist/server/MiddlewareRequest'
+import { MiddlewareResponse } from '@redwoodjs/web/dist/server/MiddlewareResponse'
 
 import type { EntryServer } from '../types'
 import { makeFilePath, ssrLoadEntryServer } from '../utils'
 
-import type { MiddlewareRequest } from './MiddlewareRequest'
-import { MiddlewareResponse } from './MiddlewareResponse'
 import type {
   Middleware,
   MiddlewareClass,
