+++ packages/vite/src/streaming/createReactStreamingHandler.ts
@@ -12,9 +12,9 @@ import type { RouteSpec, RWRouteManifestItem } from '@redwoodjs/internal'
 import { getAppRouteHook, getConfig, getPaths } from '@redwoodjs/project-config'
 import { matchPath } from '@redwoodjs/router/dist/util.js'
 import type { TagDescriptor } from '@redwoodjs/web'
+import { MiddlewareResponse } from '@redwoodjs/web/dist/server/MiddlewareResponse.js'
 
 import { invoke } from '../middleware/invokeMiddleware.js'
-import { MiddlewareResponse } from '../middleware/MiddlewareResponse.js'
 import type { Middleware } from '../middleware/types.js'
 import type { EntryServer } from '../types.js'
 import { makeFilePath, ssrLoadEntryServer } from '../utils.js'
